# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recurring:
    get:
      operationId: listRecurring
      tags: [recurring]
      summary: List recurring transactions
      responses:
        "200":
          description: List of recurring transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringTransaction"
    post:
      operationId: createRecurring
      tags: [recurring]
      summary: Create a recurring transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringTransaction"
  /recurring/{id}:
    patch:
      operationId: updateRecurring
      tags: [recurring]
      summary: Update a recurring transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringTransaction"
    delete:
      operationId: deleteRecurring
      tags: [recurring]
      summary: Delete a recurring transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /recurring/{id}/process:
    post:
      operationId: processRecurring
      tags: [recurring]
      summary: Manually trigger a recurring transaction to create its next occurrence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ─── BUDGETS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, userId, type, amount, categoryId, date, createdAt, updatedAt]
      properties:
        id:
          type: number
        userId:
          type: string
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        categoryId:
          type: number
        categoryName:
          type: ["string", "null"]
        date:
          type: string
          format: date
        note:
          type: ["string", "null"]
        recurringId:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RecurringTransaction:
      type: object
      required: [id, userId, type, amount, categoryId, recurringDay, isActive, createdAt, updatedAt]
      properties:
        id:
          type: number
        userId:
          type: string
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        categoryId:
          type: number
        categoryName:
          type: ["string", "null"]
        recurringDay:
          type: number
          minimum: 1
          maximum: 31
        note:
          type: ["string", "null"]
        isActive:
          type: boolean
        lastProcessedDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RecurringInput:
      type: object
      required: [type, amount, categoryId, recurringDay]
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          minimum: 0.01
        categoryId:
          type: number
        recurringDay:
          type: number
          minimum: 1
          maximum: 31
        note:
          type: string
    RecurringUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          minimum: 0.01
        categoryId:
          type: number
        recurringDay:
          type: number
          minimum: 1
          maximum: 31
        note:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ─── BUDGET ────────────────────────────────────────────────────────────────
...