# lib/api-zod/src/generated/types/recurringTransaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { RecurringTransactionType } from './recurringTransactionType';

export interface RecurringTransaction {
  id: number;
  userId: string;
  type: RecurringTransactionType;
  amount: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /**
     * @minimum 1
     * @maximum 31
     */
  recurringDay: number;
  /** @nullable */
  note?: string | null;
  isActive: boolean;
  /** @nullable */
  lastProcessedDate?: string | null;
  createdAt: Date;
  updatedAt: Date;
}